+++ src/azure-cli/HISTORY.rst
@@ -15,7 +15,7 @@ Release History
 
 * vm create: Add warning when specifying accelerated networking and an existing NIC together.
 * [BREAKING CHANGE] vm extension set: Fix bug where users could not set an extension on a VM with --ids.
-* New commands `az vm image terms accept/cancel/show` to manage Azure Marketplace terms.
+* New commands `az vm image terms accept/cancel/show` to manage Azure Marketplace image terms.
 
 **IoT**